+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Next release
+
+-  Record aggregate state while processing `Commanded.Aggregate.Multi` ([#507](https://github.com/commanded/commanded/pull/507)).
+
 - Properly handle EXIT signal in event handler ([#512](https://github.com/commanded/commanded/pull/512)).
 
 ## v1.4.1